+++ tests/test_cli.py
@@ -3,23 +3,12 @@ from __future__ import annotations
 import sqlite3
 import subprocess
 import sys
+from importlib.metadata import version
 from pathlib import Path
 
 import pytest
 
-from sqlacodegen.generators import _sqla_version
-
-if sys.version_info < (3, 8):
-    from importlib_metadata import version
-else:
-    from importlib.metadata import version
-
 future_imports = "from __future__ import annotations\n\n"
-
-if _sqla_version < (1, 4):
-    declarative_package = "sqlalchemy.ext.declarative"
-else:
-    declarative_package = "sqlalchemy.orm"
 
 
 @pytest.fixture
@@ -81,18 +70,19 @@ def test_cli_declarative(db_path: Path, tmp_path: Path) -> None:
 
     assert (
         output_path.read_text()
-        == f"""\
-from sqlalchemy import Column, Integer, Text
-from {declarative_package} import declarative_base
+        == """\
+from sqlalchemy import Integer, Text
+from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
-Base = declarative_base()
+class Base(DeclarativeBase):
+    pass
 
 
 class Foo(Base):
     __tablename__ = 'foo'
 
-    id = Column(Integer, primary_key=True)
-    name = Column(Text, nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    name: Mapped[str] = mapped_column(Text)
 """
     )
 
@@ -113,23 +103,19 @@ def test_cli_dataclass(db_path: Path, tmp_path: Path) -> None:
 
     assert (
         output_path.read_text()
-        == f"""\
-{future_imports}from dataclasses import dataclass, field
+        == """\
+from sqlalchemy import Integer, Text
+from sqlalchemy.orm import DeclarativeBase, Mapped, MappedAsDataclass, mapped_column
 
-from sqlalchemy import Column, Integer, Text
-from sqlalchemy.orm import registry
-
-mapper_registry = registry()
+class Base(MappedAsDataclass, DeclarativeBase):
+    pass
 
 
-@mapper_registry.mapped
-@dataclass
-class Foo:
+class Foo(Base):
     __tablename__ = 'foo'
-    __sa_dataclass_metadata_key__ = 'sa'
 
-    id: int = field(init=False, metadata={{'sa': Column(Integer, primary_key=True)}})
-    name: str = field(metadata={{'sa': Column(Text, nullable=False)}})
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    name: Mapped[str] = mapped_column(Text)
 """
     )
 
@@ -159,9 +145,75 @@ from sqlmodel import Field, SQLModel
 class Foo(SQLModel, table=True):
     id: Optional[int] = Field(default=None, sa_column=Column('id', Integer, \
 primary_key=True))
-    name: str = Field(sa_column=Column('name', Text, nullable=False))
+    name: str = Field(sa_column=Column('name', Text))
 """
     )
+
+
+def test_cli_engine_arg(db_path: Path, tmp_path: Path) -> None:
+    output_path = tmp_path / "outfile"
+    subprocess.run(
+        [
+            "sqlacodegen",
+            f"sqlite:///{db_path}",
+            "--generator",
+            "tables",
+            "--engine-arg",
+            'connect_args={"timeout": 10}',
+            "--outfile",
+            str(output_path),
+        ],
+        check=True,
+    )
+
+    assert (
+        output_path.read_text()
+        == """\
+from sqlalchemy import Column, Integer, MetaData, Table, Text
+
+metadata = MetaData()
+
+
+t_foo = Table(
+    'foo', metadata,
+    Column('id', Integer, primary_key=True),
+    Column('name', Text, nullable=False)
+)
+"""
+    )
+
+
+def test_cli_invalid_engine_arg(db_path: Path, tmp_path: Path) -> None:
+    output_path = tmp_path / "outfile"
+
+    # Expect exception:
+    # TypeError: 'this_arg_does_not_exist' is an invalid keyword argument for Connection()
+    with pytest.raises(subprocess.CalledProcessError) as exc_info:
+        subprocess.run(
+            [
+                "sqlacodegen",
+                f"sqlite:///{db_path}",
+                "--generator",
+                "tables",
+                "--engine-arg",
+                'connect_args={"this_arg_does_not_exist": 10}',
+                "--outfile",
+                str(output_path),
+            ],
+            check=True,
+            capture_output=True,
+        )
+
+    if sys.version_info < (3, 13):
+        assert (
+            "'this_arg_does_not_exist' is an invalid keyword argument"
+            in exc_info.value.stderr.decode()
+        )
+    else:
+        assert (
+            "got an unexpected keyword argument 'this_arg_does_not_exist'"
+            in exc_info.value.stderr.decode()
+        )
 
 
 def test_main() -> None: